+++ src/main/java/com/uwetrottmann/trakt5/entities/SyncEpisode.java
@@ -1,6 +1,12 @@
 package com.uwetrottmann.trakt5.entities;
 
+import com.google.gson.annotations.SerializedName;
+import com.uwetrottmann.trakt5.enums.Audio;
+import com.uwetrottmann.trakt5.enums.AudioChannels;
+import com.uwetrottmann.trakt5.enums.Hdr;
+import com.uwetrottmann.trakt5.enums.MediaType;
 import com.uwetrottmann.trakt5.enums.Rating;
+import com.uwetrottmann.trakt5.enums.Resolution;
 import org.threeten.bp.OffsetDateTime;
 
 public class SyncEpisode {
@@ -13,6 +19,13 @@ public class SyncEpisode {
     public OffsetDateTime watched_at;
     public OffsetDateTime rated_at;
     public Rating rating;
+    public MediaType media_type;
+    public Resolution resolution;
+    public Hdr hdr;
+    public Audio audio;
+    public AudioChannels audio_channels;
+    @SerializedName("3d")
+    public Boolean is3d;
 
     public SyncEpisode number(int number) {
         this.number = number;
@@ -46,6 +59,36 @@ public class SyncEpisode {
 
     public SyncEpisode rating(Rating rating) {
         this.rating = rating;
+        return this;
+    }
+
+    public SyncEpisode mediaType(MediaType media_type){
+        this.media_type = media_type;
+        return this;
+    }
+
+    public SyncEpisode resolution(Resolution resolution){
+        this.resolution = resolution;
+        return this;
+    }
+
+    public SyncEpisode hdr(Hdr hdr){
+        this.hdr = hdr;
+        return this;
+    }
+
+    public SyncEpisode audio(Audio audio){
+        this.audio = audio;
+        return this;
+    }
+
+    public SyncEpisode audioChannels(AudioChannels audio_channels){
+        this.audio_channels = audio_channels;
+        return this;
+    }
+
+    public SyncEpisode is3d(Boolean is3d){
+        this.is3d = is3d;
         return this;
     }