+++ src/main/java/com/uwetrottmann/trakt5/enums/SortHow.java
@@ -1,34 +1,13 @@
 package com.uwetrottmann.trakt5.enums;
 
-import java.util.HashMap;
-import java.util.Locale;
-import java.util.Map;
+import com.google.gson.annotations.SerializedName;
 
-public enum SortHow implements TraktEnum {
-    ASC("asc"),
-    DESC("desc");
+public enum SortHow {
 
-    private final String value;
+    @SerializedName("asc")
+    ASC,
 
-    SortHow(String value) {
-        this.value = value;
-    }
-
-    private static final Map<String, SortHow> STRING_MAPPING = new HashMap<>();
-
-    static {
-        for (SortHow via : SortHow.values()) {
-            STRING_MAPPING.put(via.toString().toUpperCase(Locale.ROOT), via);
-        }
-    }
-
-    public static SortHow fromValue(String value) {
-        return STRING_MAPPING.get(value.toUpperCase(Locale.ROOT));
-    }
-
-    @Override
-    public String toString() {
-        return value;
-    }
+    @SerializedName("desc")
+    DESC
 
 }